Simplify tile storage using bitfields
Instead of storing tiles as uint8_t, they are now stored as a
struct minesweeper_tile, which still should only be 8 bits wide,
but each property is addressable without bitmasking.

This is a breaking API change, since some functions that were
previously used to access data encoded in the tiles are no longer
needed. All data can be accessed in the tile directly.